Cypress reorganization (#71)
* Moved all the tests targeting ttd/stateless-app out of the main test folder, to allow for an easier way of running the tests against one app at a time (for when you're running locally).

* Updating package.json with new script targets for running different batches of tests against apps, or all apps. Also installing run-script-os to make this a bit more cross-platform, although I'm personally running cypress manually on linux.

* Updating and restructuring the readme with different instructions for running against a remote environment vs running entirely locally

* Tests are not being run when the tests themselves change. Updating github workflow to trigger cypress tests when there are test changes.

* Adding back if expression that I removed and committed by accident (sorry!)

* Fixing spacing to get a clean merge
