+++ .github/workflows/cypress-altinn-app-frontend.yml
@@ -6,11 +6,13 @@ on:
     paths:
       - 'src/altinn-app-frontend/**'
       - 'src/shared/**'
+      - 'test/cypress/**'
   pull_request:
     branches: [main]
     paths:
       - 'src/altinn-app-frontend/**'
       - 'src/shared/**'
+      - 'test/cypress/**'
 
   workflow_dispatch:
 
@@ -44,9 +46,7 @@ jobs:
         working-directory: test/cypress
 
       - name: Create test files
-        run: |
-          truncate -s 1K test.png
-          truncate -s 1K test.pdf
+        run: yarn run create:testfiles
         working-directory: test/cypress/e2e/fixtures
 
       - name: Verify cypress and run tests
@@ -55,7 +55,7 @@ jobs:
           export CYPRESS_PROJECT_ID=y2jhp6
           export CYPRESS_RECORD_KEY=${{ secrets.CYPRESS_RECORD_KEY }}
           export GITHUB_TOKEN=${{ secrets.GITHUB_TOKEN }}
-          ./node_modules/.bin/start-server-and-test 'cd $GITHUB_WORKSPACE/src/altinn-app-frontend && yarn start' http://localhost:8080/altinn-app-frontend.js 'yarn run test:appfrontend:headless --env environment=tt02,testUserName=tt02testuser,testUserPwd=${{ secrets.CYPRESS_ALTINN_USERPWD }} --record --parallel --tag "altinn-app-frontend" --group altinn-app-frontend --ci-build-id $GITHUB_RUN_ID-$GITHUB_RUN_ATTEMPT'
+          ./node_modules/.bin/start-server-and-test 'cd $GITHUB_WORKSPACE/src/altinn-app-frontend && yarn start' http://localhost:8080/altinn-app-frontend.js 'yarn run test:all:headless --env environment=tt02,testUserName=tt02testuser,testUserPwd=${{ secrets.CYPRESS_ALTINN_USERPWD }} --record --parallel --tag "altinn-app-frontend" --group altinn-app-frontend --ci-build-id $GITHUB_RUN_ID-$GITHUB_RUN_ATTEMPT'
         working-directory: test/cypress
 
   altinn-app-frontend-test-on-fork-pr:
@@ -82,13 +82,11 @@ jobs:
         working-directory: test/cypress
 
       - name: Create test files
-        run: |
-          truncate -s 1K test.png
-          truncate -s 1K test.pdf
+        run: yarn run create:testfiles
         working-directory: test/cypress/e2e/fixtures
 
       - name: Verify cypress and run tests
         run: |
           yarn run cy:verify
-          ./node_modules/.bin/start-server-and-test 'cd $GITHUB_WORKSPACE/src/altinn-app-frontend && yarn start' http://localhost:8080/altinn-app-frontend.js 'yarn run test:appfrontend:headless --env environment=tt02,testUserName=testuserexternal,testUserPwd=r@h74Rz7XYQJ'
+          ./node_modules/.bin/start-server-and-test 'cd $GITHUB_WORKSPACE/src/altinn-app-frontend && yarn start' http://localhost:8080/altinn-app-frontend.js 'yarn run test:all:headless --env environment=tt02,testUserName=testuserexternal,testUserPwd=r@h74Rz7XYQJ'
         working-directory: test/cypress
