+++ test/cypress/README.md
@@ -13,9 +13,10 @@ These instructions will get you run the integration tests on altinn-app-frontend
 ```
 ### Test data prerequisite
 
-For running the test against test environments like AT, TT02, test user has to be created.
+1. For running the test against remote test environments like AT, TT02, test user has to be created. You won't need
+these if you intend to only run the tests against your local setup.
 
-Create a new file name `cypress.env.json` under `test\cypress` with the data created above.
+Create a new file name `cypress.env.json` under `test\cypress` with a username and password:
 
 ```json
 {
@@ -23,48 +24,76 @@ Create a new file name `cypress.env.json` under `test\cypress` with the data cre
   "testUserPwd": ""
 }
 ```
-### Run App Frontend tests
 
-Follow the steps below to start localtest, app frontend, app and the tests.
+2. Create test files that are used by the tests as attachments in app instances. This is only needed the first time, or
+when files are deleted from `e2e/fixtures`.
 
 **Note:** Use cmd or git bash to run the scripts.
 
-1. Create testfiles that are used by the tests as attachments in app instances.
-
 ```cmd
-    yarn run create:testfiles # only needed first time, or when files are deleted from e2e/fixtures
+    yarn run create:testfiles
 ```
 
-2. Clone the app (frontend-test, stateless-app) to be tested and update config in `package.json` with the paths.
+### Running tests against a remote environment
 
-3. Start localtest, app frontend, app. (Hop over to step 4 if the solutions are already running)
+Be sure to supply credentials to `cypress.env.json` as described above.
+
+1. Start your local development server for `app-frontend-react`:
+```cmd
+    yarn run start:frontend
+```
+
+If you have the frontend dependencies installed from before, run the below command:
+
+```cmd
+    yarn run start:frontend-no-deps
+```
+
+2. Run the tests against a remote environment:
+```cmd
+    yarn run test:all -e environment=at21
+```
+
+Other remote environments could also be used (see `e2e/config/*.json`).
+
+### Running tests locally
+
+1. Clone the apps (
+[ttd/frontend-test](https://dev.altinn.studio/repos/ttd/frontend-test) and
+[ttd/stateless-app](https://dev.altinn.studio/repos/ttd/stateless-app)
+) to be tested and update config in `package.json` with the paths.
+
+2. To start localtest, app frontend, and the app you configured above, run the command below.
+(Hop over to step 3 if the solutions are already running)
 
 ```cmd
     yarn run before:appfrontend
 ```
 
-If one has the frontend dependencies installed from before, run the below command.
+If you have the frontend dependencies installed from before, run the below command:
 
 ```cmd
     yarn run before:appfrontend-no-deps
 ```
 
-4. Start the app frontend tests from a new git bash terminal.
+3. Start the tests for a given app from a new terminal:
 
 ```cmd
-    yarn run test:appfrontend -e environment=local
+    yarn run test:frontend -e environment=local
 ```
 
-To run the tests towards AT21, an altinn user credential has to be supplied [here](../cypress#test-data-prerequisite).
+or
 
 ```cmd
-    yarn run test:appfrontend -e environment=at21
+    yarn run test:stateless -e environment=local
 ```
 
-5. To run a single test case open cypress runner using
+### Running a single test
+
+To run a single test case open cypress runner using
 
 ```cmd
-    yarn run cy:open -e environment=local
+    yarn run cy:open -e environment=<environment>
 ```
 
 ### Format files with prettier
