+++ test/cypress/e2e/integration/app-stateless/stateless.js
